php-cs-fixer on all the code Created a new DefaultSetMapper interface and class for field setting methods. Removed face recognition logic from Media class that isn't on the roadmap. Added media clean up to end of tests. Model->ID is explicitly set as the primary field. Moved all getAll/get methods related to pulling records into the database into a new trait "Getters".
