+++ src/Models/Model.php
@@ -18,6 +18,8 @@ namespace Divergence\Models;
  */
 class Model extends ActiveRecord
 {
+    use Getters;
+
     /**
      * {@inheritDoc}
      */
@@ -26,6 +28,7 @@ class Model extends ActiveRecord
             'type' => 'integer',
             'autoincrement' => true,
             'unsigned' => true,
+            'primary' => true,
         ],
         'Class' => [
             'type' => 'enum',
